requirements.txt
torch==2.0.1
transformers==4.32.1
pandas==2.1.0
matplotlib==3.7.2
google-generativeai==0.2.0
selenium==4.10.0
